feat: Add ethlint support (#4)
* feat: Add ethlint support

* ci: Add contracts linter

* fix: Lint all contracts

+++ interfaces/aave/Oracle.sol
@@ -1,6 +1,7 @@
 pragma solidity ^0.5.16;
 
 interface Oracle {
-    function getAssetPrice(address reserve) external view returns (uint);
-    function latestAnswer() external view returns (uint);
+    function getAssetPrice(address reserve) external view returns (uint256);
+
+    function latestAnswer() external view returns (uint256);
 }
