+++ interfaces/dforce/Token.sol
@@ -3,8 +3,11 @@
 pragma solidity ^0.5.17;
 
 interface dERC20 {
-  function mint(address, uint256) external;
-  function redeem(address, uint) external;
-  function getTokenBalance(address) external view returns (uint);
-  function getExchangeRate() external view returns (uint);
+    function mint(address, uint256) external;
+
+    function redeem(address, uint256) external;
+
+    function getTokenBalance(address) external view returns (uint256);
+
+    function getExchangeRate() external view returns (uint256);
 }
