+++ interfaces/yearn/Converter.sol
@@ -3,5 +3,5 @@
 pragma solidity ^0.5.16;
 
 interface Converter {
-    function convert(address) external returns (uint);
+    function convert(address) external returns (uint256);
 }
