+++ interfaces/yearn/IController.sol
@@ -1,10 +1,15 @@
 pragma solidity ^0.5.16;
 
 interface IController {
-    function withdraw(address, uint) external;
-    function balanceOf(address) external view returns (uint);
-    function earn(address, uint) external;
+    function withdraw(address, uint256) external;
+
+    function balanceOf(address) external view returns (uint256);
+
+    function earn(address, uint256) external;
+
     function want(address) external view returns (address);
+
     function rewards() external view returns (address);
+
     function vaults(address) external view returns (address);
 }
