+++ interfaces/yearn/Vault.sol
@@ -1,9 +1,13 @@
 pragma solidity ^0.5.16;
 
 interface Vault {
-    function deposit(uint) external;
+    function deposit(uint256) external;
+
     function depositAll() external;
-    function withdraw(uint) external;
+
+    function withdraw(uint256) external;
+
     function withdrawAll() external;
-    function getPricePerFullShare() external view returns (uint);
+
+    function getPricePerFullShare() external view returns (uint256);
 }
